requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging & Monitoring
loguru>=0.7.0

# Data Validation
great-expectations>=0.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.26.0

# Development Tools
black>=23.11.0
flake8>=6.1.0
